requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
requests
web3
eth-account
python-telegram-bot
